Ease of use / gen 2
Added a True/False statement to allow exectuion if graphviz and or tensorflow-gpu is not installed.
Added visuilaztion of model architecture as a graph.
Added noise layer into the input in order to help with regularization

GEN2:
The future is here, here is the work in progress of this model using diffrences instead.
Grid search is used to tune hyperparameters. More to follow.